fix(credstash): replace nodecredstash with aws-credstash (#66)
`nodecredstash` depends on `aws-sdk@^2.171.0`; this elderly version of
the AWS SDK lacks support for various contemporary features.  the author
doesn't want to convert that dependency to a peerDependency in order to
facilitate parent modules enforcing a newer version (see
DavidTanner/nodecredstash#13); as a result, someone else forked
`nodecredstash` into `aws-credstash`, which is API-compatible but moves
`aws-sdk` to a peerDependency.

`aws-credstash` seems to be more actively maintained than
`nodecredstash`; this change swaps in `aws-credstash` and also enforces
a dependency on a more recent `aws-sdk` (my particular need for that
minimum version is AWS SSO credential parsing support, i would be
open to enforcing a more recent minimum version).

the test suite passes after this change, i'm about to do some additional
testing to see if i can find anything else that breaks.